# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OsaroTechStore API
  version: 1.0.0
  description: |
    REST API documentation for OsaroTechStore backend.
    This specification mirrors the current Express routes and controller behavior.
servers:
  - url: http://localhost:5000/api
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    EmptyFieldsError:
      type: object
      properties:
        error:
          type: string
        emptyFields:
          type: array
          items:
            type: string
...
    Product:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
          example: USD
        category:
          type: string
          description: Display name of the populated category.
        categoryId:
          type: string
          description: Identifier of the category used by write requests.
        stock:
          type: number
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [new, active, inactive, deprecated]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductCreateRequest:
      type: object
      required: [name, price, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
          default: USD
        category:
          type: string
        stock:
          type: number
          default: 0
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [new, active, inactive, deprecated]
          default: new
    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        category:
          type: string
        stock:
          type: number
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [new, active, inactive, deprecated]
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: Get all products
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [new, active, inactive, deprecated]
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      tags: [Products]
      summary: Create a new product
      description: Admin-only route.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - $ref: "#/components/schemas/EmptyFieldsError"
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to fetch product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Products]
      summary: Update a product
      description: Admin-only route.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Invalid id or not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Products]
      summary: Delete product by id
      description: Admin-only route.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Invalid id or not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...